pyproject.toml
[project]
name = "goustomagic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
    "requests>=2.32.3",
    "unicode>=2.9",
]
